backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Cors==4.0.0
psycopg2-binary>=2.9.0
python-dotenv==1.0.0
Authlib==1.2.1
PyJWT==2.8.0
cryptography>=41.0.0
requests==2.31.0
Werkzeug==2.3.7
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-core>=0.3.0
googlemaps>=4.10.0
